requirements.txt
fastapi
"uvicorn[standard]"
tortoise-orm[asyncpg]
